♻ Article, User 의 id 속성 타입을 long -> int 로 변경
- SimpleJdbcInsert 클래스의 executeAndReturnKey 메서드 반환 타입이 Number 이므로 불필요하게 long 으로 타입으로 변경할 필요 없이 int 타입으로 반환 받아 사용하도록 원복한다.

+++ src/main/java/com/kakao/cafe/domain/User.java
@@ -4,13 +4,13 @@ import java.util.Objects;
 
 public class User {
 
-    private long id;
+    private int id;
     private final String userId;
     private final String password;
     private final String name;
     private final String email;
 
-    public User(long id, String userId, String password, String name, String email) {
+    public User(int id, String userId, String password, String name, String email) {
         this.id = id;
         this.userId = userId;
         this.password = password;
@@ -18,11 +18,11 @@ public class User {
         this.email = email;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
